pyproject.toml
[project]
name = "law-notebooklm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.0",
    "google-genai>=1.25.0",
    "numpy>=2.3.1",
    "openai==1.55.3",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "python-docx>=0.8.11",
    "docx2txt>=0.8",
]
